pyproject.toml
[project]
name = "chatbot-teams"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "boto3>=1.42.70",
    "fastapi>=0.135.1",
    "fastmcp>=3.1.1",
    "google-api-python-client>=2.193.0",
    "langchain>=1.2.12",
    "langchain-anthropic>=1.4.0",
    "langchain-core>=1.2.19",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "python-calamine>=0.6.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "uvicorn>=0.42.0",
]
